+++ src/main/java/net/frozenblock/lib/config/v2/entry/predicates/ConfigPredicate.java
@@ -24,7 +24,7 @@ import java.util.function.Supplier;
 import com.mojang.serialization.MapCodec;
 import net.frozenblock.lib.config.v2.entry.ConfigEntry;
 import net.frozenblock.lib.config.v2.registry.ID;
-import net.frozenblock.lib.levelgen.feature.api.blockpredicates.ConfigBlockPredicate;
+import net.frozenblock.lib.levelgen.blockpredicates.ConfigBlockPredicate;
 import net.frozenblock.lib.levelgen.placement.api.ConfigPlacementFilter;
 import net.frozenblock.lib.levelgen.material.impl.ConfigConditionSource;
 import net.frozenblock.lib.item.api.loot.predicates.ConfigLootCondition;